% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say Hello to my little friend}
\usage{
hello(name)
}
\arguments{
\item{name}{Who you want to say hello to}
}
\value{
Nothing
}
\description{
Say Hello to my little friend
}
\examples{
hello("Dolly")
}
